requirements.txt
alembic==1.5.5
cffi==1.14.5
click==7.1.2
cryptography==3.4.6
Flask==1.1.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.4.4
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.4
MarkupSafe==1.1.1
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.3.23
Werkzeug==1.0.1
